models/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "models"
version = "0.1.0"
description = "A component for interacting with AI models"
dependencies = [
    "pydantic",
    "pyyaml",
    "openai",
    "lorem",
    "numpy",
    "mmh3",
    "jinja2",
    "python-dotenv",
    "aiohttp",
    "faiss-cpu",
    "langchain"
]

[project.scripts]
# No scripts needed for this library component
